Bug 1926106 - Special case _generated_background_page.html moz-extension url in nsStandardURL::EqualsInternal. r=valentin,necko-reviewers
This patch includes a special handling for the moz-extension _generated_background_page.html urls inside
nsStandardURL::EqualsInternal because it doesn't resolve to a file (instead it resolves to a data: URI,
see ExtensionProtocolHandler::ResolveSpecialCases), similarly to resource:// urls which are special case
a few lines before in the same nsStandardURL::EqualsInternal.

The test case introduced by the parent patch passes successfully on the unpacked extension loaded temporarily
with these changes applied.

Depends on D226376

Differential Revision: https://phabricator.services.mozilla.com/D226470