Bug 1927960 - Make city-based weather suggestions prefer cities nearest the client's geolocation. r=daisuke, a=dmeehan
This adds some logic on top of D226858 that computes the distance between the
client's geolocation coordinates and weather suggestions' coordinates. It works
really well. We still fall back to comparing the region and country if the
coordinates aren't present in the geolocation, but in practice, I'm not sure
Merino will ever return a geolocation without coordinates.

Also, I realized that we don't need to sort the suggestions to find the one with
the best region/country. We only need to do one loop through them.

Depends on D227271

Differential Revision: https://phabricator.services.mozilla.com/D227301